Implement domain of Review micro-service (#27)
* Add basic Hexagonal skeleton

* Add `Rating` class

* Add `Review` classes

* Configure JaCoCo to report coverage

* Update description of `Review`

Description can now be 512 characters long instead of 256.

* Update spelling of JavaDoc

* Rework differentiation between review types

This will look wierd at first, but it's basically in preperation to the
MongoDb set up. With this set up, we can get two seperate Documents
(read: tables) for the different review types. This will make it easier
to retrieve reviews of a certain type.